# doc/api.yaml
...
openapi: 3.0.4
info:
  title: WASAText API
  description: Professional API for the functionalities of the WASAText messaging app
  version: 1.0.0
...
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            description: Error response for not found resources
            properties:
              error:
                type: string
                description: Error message
                example: Not found
    UnauthorizedError:
      description: Unauthorized access
      content:
        application/json:
          schema:
            type: object
            description: Error response for unauthorized access
            properties:
              error:
                type: string
                description: Error message
                example: Unauthorized
    ValidationError:
      description: Invalid input data
      content:
        application/json:
          schema:
            type: object
            description: Error response for invalid input data
            properties:
              error:
                type: string
                description: Error message
                example: Invalid input
  schemas:
    User:
      type: object
      description: A user of the WASAText app
      required: [id, username]
      properties:
        id:
          type: integer
          description: Unique user identifier
          example: 1
        username:
          type: string
          minLength: 3
          maxLength: 32
          pattern: '^[a-zA-Z0-9_]+$'
          description: Unique username for the user
          example: johndoe
        displayName:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^.*$'
          description: Display name of the user
          example: John Doe
        profilePicture:
          type: string
          format: uri
          minLength: 10
          maxLength: 2048
          pattern: '^https?://.*$'
          description: URL of the user's profile picture
          example: https://example.com/avatar.jpg
    Reaction:
      type: object
      description: A reaction to a message
      required: [id, user, emoji]
      properties:
        id:
          type: integer
          description: Unique reaction identifier
          example: 101
        emoji:
          type: string
          minLength: 1
          maxLength: 4
          pattern: '^.*$'
          description: Emoji used for the reaction
          example: 😄
        user:
          $ref: '#/components/schemas/User'
    Message:
      type: object
      description: A message in a conversation
      required: [id, conversation_id, sender, timestamp, content, mediaType, isForwarded]
      properties:
        id:
          type: integer
          description: Unique message identifier
          example: 789
        conversation_id:
          type: integer
          description: Conversation to which this message belongs
          example: 123
        content:
          type: string
          minLength: 1
          maxLength: 4096
          pattern: '^.*$'
          description: Content of the message
          example: Hello!
        isForwarded:
          type: boolean
          description: Whether the message is forwarded
          example: true
        mediaType:
          type: string
          enum: [text, photo]
          description: Type of media in the message
        reactions:
          type: array
          minItems: 0
          maxItems: 100
          description: List of reactions to the message
          items:
            $ref: '#/components/schemas/Reaction'
        sender:
          $ref: '#/components/schemas/User'
        status:
          type: string
          enum: [sent, received, read]
          description: Status of the message
        timestamp:
          type: string
          format: date-time
          description: Time the message was sent
          example: 2025-05-30T14:48:00+00:00
...
paths:
...
  /conversations/{id}/messages:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    get:
      summary: Get messages of a conversation
      description: Retrieve all messages in a conversation
      operationId: getConversation
      tags: [message]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Messages fetched
          content:
            application/json:
              schema:
                type: array
                description: List of messages in the conversation
                minItems: 0
                maxItems: 1000
                items:
                  $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      summary: Send a new message
      description: Send a new message in a conversation
      operationId: sendMessage
      tags: [message]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: Request body for sending a new message
        content:
          application/json:
            schema:
              type: object
              description: Request body for sending a new message
              required: [content, mediaType]
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 4096
                  pattern: '^.*$'
                  description: Message content
                mediaType:
                  type: string
                  enum: [text, photo]
                  description: Type of media
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /conversations/{id}/messages/{messageId}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
      - in: path
        name: messageId
        required: true
        schema:
          type: integer
    delete:
      summary: Delete a message
      description: Delete a message from a conversation
      operationId: deleteMessage
      tags: [message]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Message deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /conversations/{id}/messages/read:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    patch:
      summary: Mark all received messages as read
      description: Marks all messages in the conversation as read for the current user.
      operationId: markMessagesRead
      tags: [message]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: All messages marked as read
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /conversations/{id}/messages/{messageId}/forward:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
      - in: path
        name: messageId
        required: true
        schema:
          type: integer
    post:
      summary: Forward a message to another conversation
      description: Forward a message to another conversation
      operationId: forwardMessage
      tags: [message]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: Request body for forwarding a message
        content:
          application/json:
            schema:
              type: object
              description: Request body for forwarding a message
              required: [targetConversationId]
              properties:
                targetConversationId:
                  type: integer
                  description: ID of the conversation to forward the message to
      responses:
        '200':
          description: Message forwarded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /conversations/{id}/messages/{messageId}/reactions:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
      - in: path
        name: messageId
        required: true
        schema:
          type: integer
    post:
      summary: Add a reaction
      description: Add a reaction to a message
      operationId: commentMessage
      tags: [message]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: Request body for adding a reaction
        content:
          application/json:
            schema:
              type: object
              description: Request body for adding a reaction
              required: [emoji]
              properties:
                emoji:
                  type: string
                  minLength: 1
                  maxLength: 4
                  pattern: '^.*$'
                  description: Emoji to add as a reaction
      responses:
        '200':
          description: Reaction added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      summary: Remove a reaction
      description: Remove a reaction from a message
      operationId: uncommentMessage
      tags: [message]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Reaction removed
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...